+++ packages/api/nestjs-nats/lib/nats-application.ts
@@ -44,12 +44,14 @@ export interface CreateSubscriberHandlerConfig {
 }
 
 export class NatsApplication {
+  private connectionManager: NatsConnectionManager
   private serviceManager: NatsServiceManager
   private subscriberManager: NatsSubscriberManager
   private consumerManager: NatsConsumerManager
   private streamManager: NatsStreamManager
 
-  constructor (private connectionManager: NatsConnectionManager) {
+  constructor () {
+    this.connectionManager = new NatsConnectionManager()
     this.serviceManager = new NatsServiceManager(this.connectionManager)
     this.subscriberManager = new NatsSubscriberManager(this.connectionManager)
     this.consumerManager = new NatsConsumerManager(this.connectionManager)
